+++ src/Fractions/Fraction.Operators.cs
@@ -13,7 +13,7 @@ public readonly partial struct Fraction {
 
     public static Fraction operator -(Fraction a, Fraction b) => a.Subtract(b);
     
-    public static Fraction operator -(Fraction a) => a.Invert();
+    public static Fraction operator -(Fraction a) => a.Negate();
 
     public static Fraction operator *(Fraction a, Fraction b) => a.Multiply(b);