+++ tests/Fractions.Tests/FractionSpecs/Negate/Method_Negate.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
 using FluentAssertions;
 using NUnit.Framework;
 using Tests.Fractions;
 
-namespace Fractions.Tests.FractionSpecs.Invert;
+namespace Fractions.Tests.FractionSpecs.Negate;
 
 [TestFixture]
 // German: Wenn ein Bruch mit 0 als Zähler und 0 als Nenner negiert wird
@@ -13,7 +12,7 @@ public class When_negating_a_fraction_with_0_as_numerator_and_0_as_denominator :
     private Fraction _result;
 
     public override void Act() {
-        _result = new Fraction(0, 0, false).Invert();
+        _result = new Fraction(0, 0, false).Negate();
     }
 
     [Test]
@@ -44,7 +43,7 @@ public class When_negating_a_fraction_with_0_as_numerator_and_4_as_denominator :
     private Fraction _result;
 
     public override void Act() {
-        _result = new Fraction(0, 4, false).Invert();
+        _result = new Fraction(0, 4, false).Negate();
     }
 
     [Test]
@@ -75,7 +74,7 @@ public class When_negating_a_fraction_with_minus_1_as_numerator_and_1_as_denomin
     private Fraction _result;
 
     public override void Act() {
-        _result = new Fraction(-1, 1, false).Invert();
+        _result = new Fraction(-1, 1, false).Negate();
     }
 
     [Test]
@@ -99,7 +98,7 @@ public class When_negating_a_fraction_with_1_as_numerator_and_1_as_denominator :
     private Fraction _result;
 
     public override void Act() {
-        _result = new Fraction(1, 1, false).Invert();
+        _result = new Fraction(1, 1, false).Negate();
     }
 
     [Test]
@@ -126,9 +125,9 @@ public class Negating_a_fraction_using_the_minus_operator : Spec {
         new Fraction(0, 4, false)
     ];
 
-    public static IEnumerable<TestCaseData> TestCases => FractionsToTest.Select(x => new TestCaseData(x).Returns(x.Invert()));
+    public static IEnumerable<TestCaseData> TestCases => FractionsToTest.Select(x => new TestCaseData(x).Returns(x.Negate()));
 
     [Test]
     [TestCaseSource(nameof(TestCases))]
-    public Fraction Is_the_same_as_calling_the_invert_method(Fraction fraction) => -fraction;
+    public Fraction Is_the_same_as_calling_the_negate_method(Fraction fraction) => -fraction;
 }